requirements.txt
langchain~=0.0.323
python-dotenv~=1.0.0
tiktoken~=0.4.0
qdrant-client==1.2.0

